\name{ATE.ncb-package}
\alias{ATE.ncb-package}
\alias{ATE.ncb}
\docType{package}
\title{
  Kernel-based covariate balancing
}
\description{
Kernel-based covariate balancing method in Wong and Chan (2018)
}
\details{
\tabular{ll}{
Package: \tab ATE.ncb\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2018-11-06\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
  Raymond Wong and Gary chan
}
\references{R. K. W. Wong and K. C. G. Chan. (2018) "Kernel-based Covariate Functional Balancing for Observational Studies". Biometrika, 105(1), 199-213.}
\examples{
}
